modified

+++ app/(admin)/dashboard/categories/components/category-form.tsx
@@ -8,34 +8,35 @@ import {
 	FormMessage,
 } from "@/components/ui/form";
 import axios from "axios";
-import { useEffect, useState } from "react";
 import toast from "react-hot-toast";
+import { TcategoryProps } from "@/types";
 import { useForm } from "react-hook-form";
+import { useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
 import { Input } from "@/components/ui/input";
 import { Loader2, Trash } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import Heading from "@/components/admin/heading";
-import getCategory from "@/actions/get-category";
 import { Separator } from "@/components/ui/separator";
 import { zodResolver } from "@hookform/resolvers/zod";
 import AlertModal from "@/components/admin/alert-modal";
 import { categoryColumnSchema, TcategoryColumnProps } from "@/schemas";
 
 export default function CategoryForm({
+	category,
 	slug,
 }: {
+	category: TcategoryProps | null;
 	slug: { id: string; new: string };
 }) {
 	const router = useRouter();
-	const categoryId = slug.id;
+	const isNewCategory = slug.id === "new";
 	const [open, setOpen] = useState(false);
-	const [categories, setcategories] = useState<TcategoryColumnProps>();
 
 	const form = useForm<TcategoryColumnProps>({
 		resolver: zodResolver(categoryColumnSchema),
-		defaultValues: categories || {
-			title: "",
+		defaultValues: {
+			title: category?.title || "",
 		},
 	});
 
@@ -44,31 +45,23 @@ export default function CategoryForm({
 	} = form;
 
 	useEffect(() => {
-		const fetchCategories = async () => {
-			try {
-				const response = await getCategory(categoryId);
-				setcategories(response.category);
-				form.reset({
-					title: response.category.title,
-				});
-			} catch (err) {
-				console.error("Error fetching initial data:", err);
-			}
-		};
-		fetchCategories();
-	}, [categoryId, form]);
+		if (category) {
+			form.reset({
+				title: category.title,
+			});
+		}
+	}, [category, form]);
 
-	const initialData = categories;
-	const action = initialData ? "Save changes" : "Create";
-	const title = initialData ? "Edit Category" : "Create Category";
-	const description = initialData ? "Edit Category" : "Add a new Category";
-	const toastMessage = initialData ? "Category updated." : "Category created.";
+	const action = category ? "Save changes" : "Create";
+	const title = category ? "Edit Category" : "Create Category";
+	const description = category ? "Edit Category" : "Add a new Category";
+	const toastMessage = category ? "Category updated." : "Category created.";
 
-	const onSubmits = async (data: TcategoryColumnProps) => {
+	const onSubmit = async (data: TcategoryColumnProps) => {
 		try {
-			if (initialData) {
-				await axios.post(
-					`https://mysticmarguerite.com/new/backend/api/category/${categoryId}`,
+			if (category && !isNewCategory) {
+				await axios.put(
+					`https://mysticmarguerite.com/new/backend/api/category/${slug.id}`,
 					data,
 				);
 			} else {
@@ -87,11 +80,13 @@ export default function CategoryForm({
 
 	const onDelete = async () => {
 		try {
-			await axios.delete(
-				`https://mysticmarguerite.com/new/backend/api/category/${categoryId}`,
-			);
-			router.push(`/dashboard/categories`);
-			toast.success("Category deleted");
+			if (!isNewCategory) {
+				await axios.delete(
+					`https://mysticmarguerite.com/new/backend/api/category/${slug.id}`,
+				);
+				router.push(`/dashboard/categories`);
+				toast.success("Category deleted");
+			}
 		} catch (error) {
 			console.error(error);
 			toast.error("Something went wrong");
@@ -113,20 +108,20 @@ export default function CategoryForm({
 					title={title}
 					description={description}
 				/>
-				{initialData && (
+				{category && !isNewCategory && (
 					<Button
 						disabled={isSubmitting}
 						variant="destructive"
 						size="sm"
 						onClick={() => setOpen(true)}>
-						<Trash className="h-4 w-4" />
+						<Trash className="h-4 w-4 text-white" />
 					</Button>
 				)}
 			</div>
 			<Separator />
 			<Form {...form}>
 				<form
-					onSubmit={form.handleSubmit(onSubmits)}
+					onSubmit={form.handleSubmit(onSubmit)}
 					className="space-y-4 w-full p-5">
 					<FormField
 						control={form.control}
